fix: use numeric datatype for amount

+++ src/main/resources/db/migration/V1__Init.sql
@@ -5,7 +5,7 @@ CREATE TABLE Report(
   org_number        INT,
   invoice_number    INT,
   party_name        VARCHAR(256),
-  amount            BIGINT,
+  amount            NUMERIC(18,4),
   currency          VARCHAR(32),
   received_at       DATE NOT NULL,
   issued_at         DATE NOT NULL
